# backend/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: pyJianYingDraft API
  description: |
    基于Django REST Framework构建的剪映草稿生成服务API
    
    ## 认证方式
    API使用JWT (JSON Web Token) 进行身份验证。
    
    ### 获取Token
    1. 使用 `/api/auth/login` 接口登录获取访问令牌
    2. 在请求头中添加: `Authorization: Bearer <access_token>`
    
    ### Token刷新
    当访问令牌过期时，使用 `/api/auth/token/refresh` 接口刷新令牌
    
  version: "2.0.0"
  contact:
    name: pyJianYingDraft
    url: https://github.com/your-repo
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api
    description: 开发环境
  - url: https://your-domain.com/api
    description: 生产环境
...
paths:
...
  /auth/token/refresh:
    post:
      tags:
        - auth
      summary: 刷新访问令牌
      description: 使用刷新令牌获取新的访问令牌
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh
              properties:
                refresh:
                  type: string
                  description: 刷新令牌
      responses:
        '200':
          description: 令牌刷新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  access:
                    type: string
                    description: 新的访问令牌
        '401':
          description: 刷新令牌无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT访问令牌
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: 错误消息
        errors:
          type: object
          description: 详细错误信息
          additionalProperties: true
...